LiteDB using has been refactored and DB size greatly diminished dashboard settings storing was implemented

+++ Perfon.Core/PerfCounters/PerformanceSumCounter.cs
@@ -27,11 +27,11 @@ namespace Perfon.Core.PerfCounters
         {
             if (ReversedPeriodValue != 0)
             {
-                return PostProcessMultiplyCoeff * Value * ReversedPeriodValue;
+                return PostProcessMultiplyCoeff * (float)Value * ReversedPeriodValue;
             }
             else
             {
-                return PostProcessMultiplyCoeff * Value;
+                return PostProcessMultiplyCoeff * (float)Value;
             }
         }
